+++ internal/portforward/automock/connection.go
