+++ internal/portforward/automock/stream.go
